Add files via upload
Two cells, first calculates pi using the monte carlo method and the second attempts to animate it. Unfortunately the plot appears blank sometimes and I have had difficulty troubleshooting it. I think the issue is related to ipympl not being updated to work with the latest matplot lib package. It will probably work if I hunt around for compatible packages or they update it.